+++ src/main/java/io/xlorey/FluxLoader/client/core/EventHandler.java
@@ -1,0 +1,7 @@
+package io.xlorey.FluxLoader.client.core;
+
+/**
+ * Client event handler
+ */
+public class EventHandler {
+}