+++ src/main/java/io/xlorey/FluxLoader/client/events/ClientEventsHandler.java
@@ -1,7 +1,0 @@
-package io.xlorey.FluxLoader.client.events;
-
-/**
- * Client event handler
- */
-public class ClientEventsHandler {
-}