+++ src/main/java/io/xlorey/FluxLoader/server/core/Core.java
@@ -1,6 +1,5 @@
 package io.xlorey.FluxLoader.server.core;
 
-import io.xlorey.FluxLoader.server.events.ServerEventsHandler;
 import io.xlorey.FluxLoader.shared.EventManager;
 import io.xlorey.FluxLoader.shared.PluginManager;
 import io.xlorey.FluxLoader.utils.Logger;
@@ -18,7 +17,7 @@ public class Core {
 
         Logger.print("FluxLoader Core initialization for the server..");
 
-        EventManager.subscribe(new ServerEventsHandler());
+        EventManager.subscribe(new EventsHandler());
 
         PluginManager.loadPlugins(false);
     }